adding delete file tests and refactor old tests
