+++ tools/integration_tests/operations/file_attributes_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Provides integration tests for file operations.
+// Provides integration tests for file attributes.
 package operations_test
 
 import (
@@ -21,27 +21,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/operations"
 	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/setup"
 )
-
-func TestRenameFile(t *testing.T) {
-	fileName := setup.CreateTempFile()
-
-	content, err := operations.ReadFile(fileName)
-	if err != nil {
-		t.Errorf("Read: %v", err)
-	}
-
-	newFileName := fileName + "Rename"
-
-	err = operations.RenameFile(fileName, newFileName)
-	if err != nil {
-		t.Errorf("Error in file copying: %v", err)
-	}
-	// Check if the data in the file is the same after renaming.
-	setup.CompareFileContents(t, newFileName, string(content))
-}
 
 func TestFileAttributes(t *testing.T) {
 	preCreateTime := time.Now()
@@ -64,28 +45,4 @@ func TestFileAttributes(t *testing.T) {
 	if fStat.Size() != 14 {
 		t.Errorf("File size is not 14 bytes, found size: %d bytes", fStat.Size())
 	}
-}
-
-func TestCopyFile(t *testing.T) {
-	fileName := setup.CreateTempFile()
-
-	content, err := operations.ReadFile(fileName)
-	if err != nil {
-		t.Errorf("Read: %v", err)
-	}
-
-	newFileName := fileName + "Copy"
-	if _, err := os.Stat(newFileName); err == nil {
-		t.Errorf("Copied file %s already present", newFileName)
-	}
-
-	err = operations.CopyFile(fileName, newFileName)
-	if err != nil {
-		t.Errorf("Error : %v", err)
-	}
-
-	// Check if the data in the copied file matches the original file,
-	// and the data in original file is unchanged.
-	setup.CompareFileContents(t, newFileName, string(content))
-	setup.CompareFileContents(t, fileName, string(content))
 }
